requirements.txt
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
matplotlib>=3.5.0
pandas>=1.3.0
